+++ src/entity/helpers.js
@@ -1,4 +1,4 @@
-// import isDate from 'lodash/isDate'
+import isDate from 'lodash/isDate'
 import isEmpty from 'lodash/isEmpty'
 import isPlainObject from 'lodash/isPlainObject'
 import isString from 'lodash/isString'
@@ -8,20 +8,24 @@ export function nextId() {
   return Math.random().toString(36).substr(6)
   .substring(1, 9)
 }
-// Add fields required for save.
-export function insertFields(data) {
-  return {
-    ...data,
-    dateCreated: new Date(),
-    id: nextId(),
-  }
-}
 // Check to see if object has required fields.
 export function isEntity(value) {
-  return isPlainObject(value) && value.type && isString(value.type)
+  return isPlainObject(value) && isString(value.type)
 }
+
 // Check to see if the entity has been populated with fields required for save.
-export function isEntityCreated(entity) {
-  return !isEmpty(entity.id)
-  // && isDate(entity.dateCreated)
+export function isEntityCreated(entity, checkDate = false) {
+  if (!isEntity(entity) || isEmpty(entity.id)) return false
+  if (checkDate && !isDate(entity.dateCreated)) return false
+  return true
+}
+// Add fields required for save.
+export function insertFields(data) {
+  if (isEntityCreated(data, true)) return data
+  return {
+    ...data,
+    dateCreated: data.dateCreated || new Date(),
+    id: data.id || nextId(),
+    type: data.type || 'Thing',
+  }
 }