+++ src/index.js
@@ -3,7 +3,7 @@ import entity from './entity/reducer'
 import triple from './triple/reducer'
 
 export { create, createIfNew, createTriple } from './helpers'
-export { isEntity, isEntityCreated, nextId } from './entity/helpers'
+export { insertFields, isEntity, isEntityCreated, nextId } from './entity/helpers'
 export { isTriple } from './triple/helpers'
 
 export {